style: fixed responsivity issues
Flexbox has a nasty issue when it sets the default `min-width` value to `auto`. This causes a weird behavior when there's a flex child with `white-space: nowrap`, even if it's deeply nested into the DOM tree, causing all children up until the outer-most flex container to have a width larger than the parent if the content is large enough.