refactor: changed getTranslationDeposit implementation
Instead of adding a value porportional to the current translator deposit as fail-safe mechanism for transaction mining delay, we now consider the pricing function growth to estimate a safe value for the deposit.

This way we don't overestimate the value for tasks whose payout increase
very slowly and don't underestimate it when it increases too fast.