fix: wrong result for resolved tasks
When a task is resolved without a dispute, the `ruling` parameter is never set, which meant it would stick with the default `0` value. This was causing approved tasks to be shown as if there was a dispute, but the jurors refused to rule.

To fix this problem, we set the `ruling` attribute of a dispute-less task that is in `Resolved` status through the `reason` parameter in the `TaskResolved` event.