refactor(tokens): allow reading information from unsupported tokens
This change deals with a problem of a token being listed on T2CR in a point of time, but being removed after that.

When this happened, Linguo was not able to display the token symbol on the UI and would show a "???" mark.

Now information for tokens is automatically pulled directly from the smartcontract even if they are not present in T2CR anymore.