fix(#18): fixed upload for non-ASCII characters in file name
Currently the Kleros IPFS proxy does not support file names with non-ASCII characters, making it crash when trying to visualize the uploaded fiel.

To prevent that, we will generate a determnistic ASCII-only name for file names containing non-ASCII characters.

There was also an issue with file upload charset, which the user would only notice after submiting a transaction with the file path. We added the ability to visualize the file right after the upload.