refactor: inverted dependency relation to fetch tasks after create
Instead of making the tasks slice (more general) to depend upon the requester slice (more specialized) by emitting a special requester/fetchTasks action, the tasks slice now emits an event when the transaction that creates the task is mined. The requester slice listens to such action and dispatch the respective requester/fetchTask.

This also fixes a long-lasting problem of not being able to fetch tasks right after the transaction was mined. To avoid reading from all contracts for all users, we use the Etherscan API to see which contracts the user has interacted with in the past 60 days to determine which contracts we will query for tasks. However, this scheme would break if the user was interacting with Linguo for the first time or if she was requesting a translation for a language group pair she never requested before, since each language pair is a different contract and it might take a while for that interaction to show up in Etherscan API results.

To avoid that, we changed the implementation of the Linguo API facade to allow the consumers to provide a hint of language group pairs to be included in the query. Since the created task params are included in the action payload, we can infer the language group pair and provide the hint to allow fetching the new task right the way.

PHEW.