fix: added transient notifications handling on refresh
Currently transient notifications for resovled tasks could fail to be updated to the final state in case there was a page refresh during the task details fetching. This would cause the notification to be in the transient state forever.

We introduced a way to fetch all transient notifications and re-process them.