perf: fixed always fetching data from older blocks on tasks update
Previously the latest block was only updated if a notification for the user was created. However this caused the same blocks to be queried over and over again whenever the user visited the page, which would cause a slowdown.

We changed the ownership of the latest block information to the task updates subslice, which will update it after receiving any kind of event from the contracts.