fix: fixed open tasks filter for translator
There's no such thing as my open tasks. When showing only open tasks, users should see all tasks by default.

The catch is that we don't want to mess with the previous selected allTasks filter, so when the user selects another status, it preserves its previous value.

To do that, we need to change the filter params in a place where it does not affect the store.